requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
vaderSentiment>=3.3.2
google-api-python-client>=2.100.0
pytest>=7.0.0
httpx>=0.26.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.13.0
redis>=5.0.0
prometheus-fastapi-instrumentator>=7.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart
slowapi>=0.1.9
gunicorn>=21.2.0
